Replaces descriptor sets with ShaderBindingContext
Removes direct usage of DescriptorSets and integrates ShaderBindingContext for managing resource bindings.

This change simplifies resource management, improves code readability, and enables more flexible and efficient binding of resources in shaders. It replaces explicit descriptor set writes and binds with ShaderBindingContext calls.

+++ src/Cory/Renderer/Common.hpp
@@ -55,9 +55,9 @@ using CompilationResult = std::expected<ShaderCompilationOutput, CompilationErro
 
 using BufferDeviceAddress = uint64_t;
 
-using TextureHeapIndex = uint16_t;
-using BufferHeapIndex = uint16_t;
-using SamplerHeapIndex = uint16_t;
+using TextureHeapIndex = uint32_t;
+using BufferHeapIndex = uint32_t;
+using SamplerHeapIndex = uint32_t;
 
 // enums
 static constexpr Gpu::ShaderStageFlagBits SHADER_TYPE_UNKNOWN = Gpu::ShaderStageFlagBits{};
